# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Omni - Documentação da API",
        "description": "Documentação de todos os endpoints da API do sistema Omni.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/admin/projects": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Lista todos os projetos",
                "operationId": "c1725e676a2c746e33b7655c04753e65",
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Project"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Projects"
                ],
                "summary": "Cria um novo projeto",
                "operationId": "0b4e209628531c3d1e5bb311f33eae3d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreProjectRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Project"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/projects/{id}": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Busca um projeto pelo ID",
                "operationId": "0402daefb5f33d9b093b69c31c214eeb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Project"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Projeto não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Projects"
                ],
                "summary": "Atualiza um projeto",
                "operationId": "84252f77768bb1c5ddbec6c5a83bbfed",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateProjectRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Project"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Projeto não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Projects"
                ],
                "summary": "Deleta um projeto",
                "operationId": "27b2bd54546aec83fa51ef03cfc604b9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deletado com sucesso"
                    },
                    "404": {
                        "description": "Projeto não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/projects-user": {
            "get": {
                "tags": [
                    "User Projects"
                ],
                "summary": "Lista todas as associações entre usuários e projetos",
                "operationId": "7acabf2de1bf66dce0044418d4f5027e",
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ProjectUser"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "User Projects"
                ],
                "summary": "Cria uma nova associação entre usuário e projeto",
                "operationId": "874f99848eee9944c0103cb9dea91bb9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreProjectUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectUser"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/projects-user/{id}": {
            "get": {
                "tags": [
                    "User Projects"
                ],
                "summary": "Busca uma associação pelo ID",
                "operationId": "007a57dcf3251970fd505d07d4f4c84c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectUser"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Associação não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "User Projects"
                ],
                "summary": "Atualiza uma associação (não comum, geralmente se deleta e cria)",
                "operationId": "87fa88ce2c69e2c2cd62066914433aaa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreProjectUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectUser"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Associação não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "User Projects"
                ],
                "summary": "Deleta uma associação entre usuário e projeto",
                "operationId": "33abfd75d38a2d71a8825ea807431151",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deletado com sucesso"
                    },
                    "404": {
                        "description": "Associação não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreProjectRequest": {
                "title": "Store Project Request",
                "required": [
                    "client_id",
                    "name",
                    "status",
                    "type_project",
                    "date_project"
                ],
                "properties": {
                    "client_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Novo Projeto de Website"
                    },
                    "status": {
                        "description": "Status do projeto",
                        "type": "string",
                        "enum": [
                            "not_started",
                            "raw",
                            "started",
                            "in_progress",
                            "awaiting_approval",
                            "completed",
                            "postponed",
                            "canceled"
                        ],
                        "example": "not_started"
                    },
                    "type_project": {
                        "type": "string",
                        "example": "Desenvolvimento"
                    },
                    "date_project": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-12-31"
                    },
                    "visibility": {
                        "type": "string",
                        "enum": [
                            "public",
                            "private"
                        ],
                        "example": "private",
                        "nullable": true
                    },
                    "members": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "example": [
                            1,
                            2
                        ],
                        "nullable": true
                    },
                    "description": {
                        "type": "string",
                        "example": "Descrição detalhada do projeto.",
                        "nullable": true
                    },
                    "link": {
                        "type": "string",
                        "format": "url",
                        "example": "https://projeto.com",
                        "nullable": true
                    },
                    "goal": {
                        "type": "string",
                        "example": "Aumentar as vendas em 20%",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "UpdateProjectRequest": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Projeto de Marketing Atualizado"
                    },
                    "status": {
                        "type": "string",
                        "example": "completed"
                    },
                    "visibility": {
                        "type": "string",
                        "enum": [
                            "public",
                            "private"
                        ],
                        "example": "public"
                    },
                    "members": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "example": [
                            1,
                            3
                        ]
                    }
                },
                "type": "object"
            },
            "StoreProjectUserRequest": {
                "required": [
                    "user_id",
                    "project_id"
                ],
                "properties": {
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "project_id": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
            "Column": {
                "title": "Column",
                "description": "Representa uma coluna (ou estágio) dentro de um projeto Kanban.",
                "properties": {
                    "id": {
                        "description": "ID único da coluna",
                        "type": "integer",
                        "readOnly": true
                    },
                    "project_id": {
                        "description": "ID do projeto ao qual a coluna pertence",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Nome da coluna",
                        "type": "string",
                        "example": "A Fazer"
                    },
                    "created_at": {
                        "description": "Data de criação",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updated_at": {
                        "description": "Data da última atualização",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    }
                },
                "type": "object"
            },
...
            "Project": {
                "title": "Project",
                "description": "Representa um projeto, que contém colunas, tarefas e membros.",
                "properties": {
                    "id": {
                        "description": "ID único do projeto",
                        "type": "integer",
                        "readOnly": true
                    },
                    "client_id": {
                        "description": "ID do cliente associado ao projeto",
                        "type": "integer"
                    },
                    "user_id": {
                        "description": "ID do usuário que criou o projeto",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Nome do projeto",
                        "type": "string",
                        "example": "Lançamento Campanha de Verão"
                    },
                    "description": {
                        "description": "Descrição detalhada do projeto",
                        "type": "string",
                        "nullable": true
                    },
                    "status": {
                        "description": "Status atual do projeto",
                        "type": "string",
                        "enum": [
                            "pending",
                            "in_progress",
                            "completed",
                            "canceled"
                        ],
                        "example": "in_progress"
                    },
                    "visibility": {
                        "description": "Visibilidade do projeto",
                        "type": "string",
                        "enum": [
                            "public",
                            "private"
                        ],
                        "example": "private"
                    },
                    "type_project": {
                        "description": "Tipo ou natureza do projeto",
                        "type": "string",
                        "example": "Marketing Digital",
                        "nullable": true
                    },
                    "link": {
                        "description": "Link externo relacionado ao projeto",
                        "type": "string",
                        "format": "uri",
                        "example": "http://miro.com/board/12345",
                        "nullable": true
                    },
                    "goal": {
                        "description": "O objetivo principal do projeto",
                        "type": "string",
                        "nullable": true
                    },
                    "date_project": {
                        "description": "Data de início ou entrega do projeto",
                        "type": "string",
                        "format": "date",
                        "example": "2025-12-31"
                    },
                    "category": {
                        "description": "Categoria do projeto",
                        "type": "string",
                        "example": "E-commerce",
                        "nullable": true
                    },
                    "segment": {
                        "description": "Segmento de mercado do projeto",
                        "type": "string",
                        "example": "Moda",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Data de criação",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updated_at": {
                        "description": "Data da última atualização",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "columns": {
                        "description": "Lista de colunas do projeto",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Column"
                        },
                        "readOnly": true
                    },
                    "members": {
                        "description": "Usuários que têm acesso ao projeto",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/User"
                        },
                        "readOnly": true
                    }
                },
                "type": "object"
            },
            "ProjectUser": {
                "title": "User Project Association",
                "description": "Modelo que representa a ligação entre um usuário e um projeto",
                "properties": {
                    "id": {
                        "description": "ID único da associação",
                        "type": "integer",
                        "readOnly": true
                    },
                    "user_id": {
                        "description": "ID do usuário associado",
                        "type": "integer",
                        "example": 1
                    },
                    "project_id": {
                        "description": "ID do projeto associado",
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "description": "Data de criação da associação",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updated_at": {
                        "description": "Data da última atualização da associação",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}